添加地块，添加callback

+++ app/src/main/java/com/xp/zjd/photo/PhotosFragment.java
@@ -29,6 +29,7 @@ import com.loopj.android.http.BinaryHttpResponseHandler;
 import com.loopj.android.image.SmartImage;
 import com.loopj.android.image.SmartImageView;
 import com.xp.R;
+import com.xp.common.po.Status;
 import com.xp.common.tools.ImageViewTouch;
 import com.xp.common.tools.OkHttpClientUtils;
 import com.xp.zjd.po.ZJD;
@@ -103,39 +104,43 @@ public class PhotosFragment extends Fragment {
         @Override
         public void onClick(View v) {
 
-                // 创建File对象，用于存储拍照后的图片
-                //存放在手机SD卡的应用关联缓存目录下
-                File outputImage = new File(PhotoService.dirRoot, "临时照片.jpg");
+            if (Tool.isEmpty(zjd.getZDNUM())) {
+                AndroidTool.showAnsyTost("没有宗地编码，不能拍照", Status.Error);
+                return;
+            }
+            // 创建File对象，用于存储拍照后的图片
+            //存放在手机SD卡的应用关联缓存目录下
+            File outputImage = new File(PhotoService.dirRoot, "临时照片.jpg");
 
-                Tool.exitsDir(outputImage.getParent(), true);
-                // 从Android 6.0系统开始，读写SD卡被列为了危险权限，如果将图片存放在SD卡的任何其他目录，
-                //   都要进行运行时权限处理才行，而使用应用关联 目录则可以跳过这一步
-                try {
-                    if (outputImage.exists()) {
-                        outputImage.delete();
-                    }
-                } catch (Exception e) {
-                    e.printStackTrace();
+            Tool.exitsDir(outputImage.getParent(), true);
+            // 从Android 6.0系统开始，读写SD卡被列为了危险权限，如果将图片存放在SD卡的任何其他目录，
+            //   都要进行运行时权限处理才行，而使用应用关联 目录则可以跳过这一步
+            try {
+                if (outputImage.exists()) {
+                    outputImage.delete();
                 }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
 
                   /* 7.0系统开始，直接使用本地真实路径的Uri被认为是不安全的，会抛 出一个FileUriExposedException异常。
                    而FileProvider则是一种特殊的内容提供器，它使用了和内 容提供器类似的机制来对数据进行保护，
                    可以选择性地将封装过的Uri共享给外部，从而提高了 应用的安全性*/
 
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                    //大于等于版本24（7.0）的场合
-                    imageUri = FileProvider.getUriForFile(view.getContext(), "com.xp.fileprovider", outputImage);
-                } else {
-                    //小于android 版本7.0（24）的场合
-                    imageUri = Uri.fromFile(outputImage);
-                }
-                //启动相机程序
-                Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);//
-                startActivityForResult(intent, TAKE_CAMERA);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                //大于等于版本24（7.0）的场合
+                imageUri = FileProvider.getUriForFile(view.getContext(), "com.xp.fileprovider", outputImage);
+            } else {
+                //小于android 版本7.0（24）的场合
+                imageUri = Uri.fromFile(outputImage);
+            }
+            //启动相机程序
+            Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+            intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);//
+            startActivityForResult(intent, TAKE_CAMERA);
             try {
-            }catch (Exception e){
-                Toast.makeText(AndroidTool.getMainActivity(),e.getMessage(),Toast.LENGTH_LONG).show();
+            } catch (Exception e) {
+                Toast.makeText(AndroidTool.getMainActivity(), e.getMessage(), Toast.LENGTH_LONG).show();
             }
 
         }
@@ -144,22 +149,22 @@ public class PhotosFragment extends Fragment {
     @Override
     public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
 
-            switch (requestCode) {
-                case TAKE_CAMERA:
-                    if (resultCode == RESULT_OK) {
-                        // 将拍摄的照片显示出来
-                        String path = PhotoService.dirRoot + "临时照片.jpg";
-                        Photo photo = new Photo(path, false);
-                        editName(photo, 0);
+        switch (requestCode) {
+            case TAKE_CAMERA:
+                if (resultCode == RESULT_OK) {
+                    // 将拍摄的照片显示出来
+                    String path = PhotoService.dirRoot + "临时照片.jpg";
+                    Photo photo = new Photo(path, false);
+                    editName(photo, 0);
 
-                    }
-                    break;
-                default:
-                    break;
-            }
+                }
+                break;
+            default:
+                break;
+        }
         try {
-        }catch (Exception e){
-            Toast.makeText(AndroidTool.getMainActivity(),e.getMessage(),Toast.LENGTH_LONG).show();
+        } catch (Exception e) {
+            Toast.makeText(AndroidTool.getMainActivity(), e.getMessage(), Toast.LENGTH_LONG).show();
         }
 
     }
@@ -334,7 +339,6 @@ public class PhotosFragment extends Fragment {
             return convertView;
 
         }
-
 
 
         /**
